+++ src/app/map/compass.ts
@@ -1,0 +1,11 @@
+
+export interface CompassHeading {
+    magneticHeading: number, 
+    trueHeading: number, 
+    headingAccuracy: number, 
+    timestamp: number
+}
+
+export interface CompassError {
+    code: any
+}